requirements.txt
asgiref==3.8.1
Django==5.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
django-jazzmin==3.0.1
django-js-asset==2.2.0